requirements.txt
graphviz==0.20.1
openai>=0.27.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
itsdangerous>=2.1.2
authlib>=1.2.0
python-dotenv>=1.0.0
